backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
gtts==2.4.0
faiss-cpu==1.7.4
pandas==2.1.4
aiofiles==23.2.1
